+++ front-end/.gitignore
@@ -9,6 +9,7 @@ npm-debug.*
 *.orig.*
 web-build/
 web-report/
+auth.json
 
 # macOS
 .DS_Store
